backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
